test(crypto): cover ERC-20 transferFrom gas-attribution case
Locks in coverage for a router (or any third-party contract holding
prior approval) pulling tokens out of our wallet via `transferFrom`.
The transfer row's `from` is the wallet, but the EOA that signed the
outer tx — and paid the gas — is the router caller. The gating
predicate added in Task 2 already covers this; this test pins it.

Splits the "Outer-tx signed by someone else → no gas leg" MARK block
out of `TransferEventBuilderGasLegTests` into a sibling
`TransferEventBuilderGasAttributionTests` suite so each suite stays
inside SwiftLint's `type_body_length` budget — same precedent as the
earlier `TransferEventBuilderGasCoalescingTests` split. The Task 2
internal-transfer test moves alongside its sibling unchanged.
