+++ MoolahTests/Shared/CryptoImport/TransferEventBuilderGasLegTests.swift
@@ -248,56 +248,4 @@ struct TransferEventBuilderGasLegTests {
     #expect(badCandidate.transaction.legs.allSatisfy { $0.type == .expense })
   }
 
-  // MARK: - Outer-tx signed by someone else → no gas leg
-
-  @Test(
-    "Internal transfer where wallet appears as `from` but receipt.from is counterparty → no gas leg"
-  )
-  func internalTransferSignedBySomeoneElseDoesNotEmitGasLeg() async throws {
-    let subject = makeDiscoverySubject()
-    let alchemy = RecordingAlchemyClientStub()
-    // Receipt is fetched (the outboundHashes heuristic still trips on
-    // `from == wallet` for any non-NFT transfer), but the EOA that
-    // signed the tx is the counterparty — wallet did not pay gas.
-    alchemy.setReceiptResponse(
-      .receipt(
-        AlchemyTransactionReceipt(
-          hash: "0xinternal",
-          gasUsed: Self.gasUsed,
-          effectiveGasPrice: Self.gasPrice,
-          from: Self.counterparty)
-      ),
-      for: "0xinternal")
-
-    let account = makeCryptoAccount(walletAddress: Self.wallet, chain: .ethereum)
-    let origin = makeWalletImportOrigin(for: account.id)
-    // Wallet is `from` of an `internal` sub-call inside someone else's
-    // outer tx. Real example: a contract Alice called moves wallet's
-    // funds via a prior approval / delegate.
-    let transfer = makeAlchemyTransfer(
-      hash: "0xinternal",
-      from: Self.wallet,
-      to: Self.counterparty,
-      category: .internal)
-
-    let built = try await TransferEventBuilder().build(
-      transfers: [transfer],
-      account: account,
-      services: BuilderServices(
-        chain: .ethereum,
-        discovery: subject.service,
-        alchemy: alchemy),
-      importOrigin: origin)
-
-    let candidate = try #require(built.first)
-    // Transfer leg still produced; gas leg is suppressed.
-    #expect(candidate.transaction.legs.count == 1)
-    #expect(
-      candidate.transaction.legs.allSatisfy {
-        $0.externalId?.hasSuffix(":gas") == false
-      })
-    // The receipt fetch still happened — gating is post-fetch, by design.
-    #expect(alchemy.recordedReceiptCalls == ["0xinternal"])
-  }
-
 }
